Revert "CM-1326 backup permissions (#271)"
This reverts commit 4f33605133807a163ec4ffbcf388d1c2cc601b49.
