+++ functions/src/constants/index.ts
@@ -40,10 +40,6 @@ interface Env {
     notifications: string;
   }
 
-  storage: {
-    bucket: string;
-  }
-
   secretManagerProject: string;
   local: string;
