+++ functions/src/crons/backupCron.ts
@@ -1,13 +1,13 @@
 import * as functions from 'firebase-functions';
 
-import { backup } from '../util/backup';
+import backupUtil from '../util/backup';
 
-export const backupCron = functions.pubsub
+export const backup = functions.pubsub
   .schedule('0 */3 * * *')
   .onRun(async () => {
     logger.info('🚀 Beginning backup procedure');
 
-    await backup();
+    await backupUtil.backup();
 
     logger.info('✨ Backup procedure done successfully');
   });
