+++ functions/src/crons/index.ts
@@ -1,1 +1,1 @@
-export { backupCron } from './backupCron';
+export { backup } from './backupCron';
